[#5483]: Add safeguard to creating/deleting nodes in a universe with conflicting names.
Summary:
Currently, if there are two yugaware's using the same AWS configs, and they create
identically named universes, the YW that created the universe first succeeds while the second one
fails due to the node name being taken. But, when deleting the failed universe, YW will go ahead and
delete the nodes belonging to the other universe since it queries the instances by name. This diff
addresses this by preventing deletion of nodes unless they have their private IPs assigned, since
that field is only populated on a successful provisioning of the instance. This diff also adds a
safeguard in the universe creation itself, in which if there was a mismatch in the number of nodes
of the two universes in the different YWs, the new YW can execute tasks to create the nodes that
don't exist while failing on the other ones.

Test Plan:
Created AWS/GCP universes and then manually deleted them from the postgres DB. The tried
to re-create the universe and verified that it failed as expected. Then ran a delete command and
checked that it logged that it skipped all nodes that it did not have IPs for, but went ahead and
removed the universe from YW.

Reviewers: sanketh, daniel

Reviewed By: daniel

Subscribers: spotachev, streddy, jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D9765

+++ managed/devops/opscli/ybops/cloud/gcp/cloud.py
@@ -73,6 +73,10 @@ class GcpCloud(AbstractCloud):
         self.get_admin().mount_disk(args.zone, instance, body)
 
     def delete_instance(self, args):
+        host_info = self.get_host_info(args)
+        if args.node_ip is None or host_info['private_ip'] != args.node_ip:
+            logging.error("Host {} IP does not match.".format(args.search_pattern))
+            return
         self.get_admin().delete_instance(args.zone, args.search_pattern)
 
     def get_regions(self, args):
